feat: 移除智能联想插件并更新版本至2.0.0-beta1
- 移除prediction-onnx智能联想插件相关代码和配置
- 从构建工作流中删除ONNX模型文件下载和验证步骤
- 从ExtensionManager中移除PredictionPlugin和SpeechPlugin相关API
- 更新应用版本号从1.4.0到2.0.0-beta1
- 调整README.md功能描述，替换为语音转文本和表情插件说明
- 优化ProGuard规则配置
- 移除相关的测试用例和文档文件
