+++ README.md
@@ -35,8 +35,8 @@
 
 - **五笔86输入方案** - 支持五笔86及五笔拼音混输方案
 - **Rime 引擎** - 使用成熟稳定的 Rime 输入法引擎
-- **智能联想** - 基于 ONNX 模型的智能联想词推荐
-- **个性化学习** - 自动学习用户输入习惯，提升联想准确度
+- **语音转文本** - 内置语音识别功能（支持阿里百炼 FunAsr）
+- **表情插件** - 支持扩展表情插件（颜文字、表情包等）
 - **简洁界面** - Material Design 3 风格，支持浅色/深色主题
 - **主题定制** - 多种键盘配色方案可选
 - **按键反馈** - 可调节音效和振动强度
