+++ plugins/emoji-sticker/build.gradle.kts
@@ -20,12 +20,8 @@ android {
 
     buildTypes {
         release {
-            isMinifyEnabled = true
+            isMinifyEnabled = false
             isShrinkResources = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
         }
     }
     
@@ -57,14 +53,9 @@ android.applicationVariants.all {
 }
 
 dependencies {
-    // Force annotations version to resolve conflict
     constraints {
         implementation("org.jetbrains:annotations:23.0.0")
     }
-    
-    // All dependencies from plugin-core (api) + host app
     compileOnly(project(":plugin-core"))
-    
-    // Plugin-specific dependency (must be implementation)
     implementation("io.coil-kt:coil-compose:2.5.0")
 }
