+++ plugins/emoji-sticker/proguard-rules.pro
@@ -1,25 +1,9 @@
 # Plugin ProGuard rules
-# Disable obfuscation
 -dontobfuscate
-
-# CRITICAL: Only disable Lambda merging optimization
 -optimizations !class/merging/*
 
-# Keep Kotlin stdlib
 -keep class kotlin.** { *; }
--keepclassmembers class kotlin.** { *; }
+-keepnames class kotlin.** { *; }
 
-# Keep Compose classes
--keep class androidx.compose.** { *; }
--keepclassmembers class androidx.compose.** { *; }
-
-# Keep kotlinx.coroutines
--keep class kotlinx.coroutines.** { *; }
--keepclassmembers class kotlinx.coroutines.** { *; }
-
-# Keep plugin classes
 -keep class com.kingzcheung.kime.plugin.emoji.** { *; }
--keepclassmembers class com.kingzcheung.kime.plugin.emoji.** { *; }
-
-# Preserve line numbers
 -keepattributes SourceFile,LineNumberTable
