--- plugins/emoji-sticker/src/main/java/com/kingzcheung/kime/plugin/emoji/PluginSettingsActivity.kt
@@ -1,42 +1,0 @@
-package com.kingzcheung.kime.plugin.emoji
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.material3.*
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.graphics.Color
-import com.kingzcheung.kime.plugin.emoji.ui.EmojiSettingsScreen
-
-class PluginSettingsActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        
-        setContent {
-            val darkTheme = isSystemInDarkTheme()
-            
-            MaterialTheme(
-                colorScheme = if (darkTheme) {
-                    darkColorScheme(
-                        primary = Color(0xFF81D4FA),
-                        secondary = Color(0xFF81D4FA),
-                        tertiary = Color(0xFF81D4FA)
-                    )
-                } else {
-                    lightColorScheme(
-                        primary = Color(0xFFFF6B6B),
-                        secondary = Color(0xFFFF6B6B),
-                        tertiary = Color(0xFFFF6B6B)
-                    )
-                }
-            ) {
-                EmojiSettingsScreen(
-                    onNavigateBack = { finish() }
-                )
-            }
-        }
-    }
-}
