--- plugins/emoji-sticker/src/main/java/com/kingzcheung/kime/plugin/emoji/ui/EmojiSettingsScreen.kt
@@ -1,178 +1,0 @@
-package com.kingzcheung.kime.plugin.emoji.ui
-
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.items
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.Face
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import coil.compose.AsyncImage
-import coil.request.ImageRequest
-import java.io.File
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun EmojiSettingsScreen(
-    onNavigateBack: () -> Unit
-) {
-    val context = LocalContext.current
-    
-    var emojiFiles by remember { mutableStateOf<List<File>>(emptyList()) }
-    var isLoading by remember { mutableStateOf(true) }
-    
-    LaunchedEffect(Unit) {
-        val emojisDir = File(context.filesDir, "emojis")
-        if (!emojisDir.exists()) {
-            val assetManager = context.assets
-            val emojiAssetFiles = assetManager.list("emojis") ?: emptyArray()
-            
-            emojisDir.mkdirs()
-            for (fileName in emojiAssetFiles) {
-                assetManager.open("emojis/$fileName").use { input ->
-                    File(emojisDir, fileName).outputStream().use { output ->
-                        input.copyTo(output)
-                    }
-                }
-            }
-        }
-        
-        emojiFiles = emojisDir.listFiles()
-            ?.filter { it.extension in listOf("jpg", "png", "gif") }
-            ?.sortedBy { it.nameWithoutExtension.toIntOrNull() ?: 0 }
-            ?: emptyList()
-        
-        isLoading = false
-    }
-    
-    Scaffold(
-        containerColor = MaterialTheme.colorScheme.background,
-        topBar = {
-            TopAppBar(
-                title = { 
-                    Text(
-                        "恶搞兔表情包",
-                        style = MaterialTheme.typography.titleMedium
-                    )
-                },
-                navigationIcon = {
-                    IconButton(onClick = onNavigateBack) {
-                        Icon(
-                            Icons.AutoMirrored.Filled.ArrowBack,
-                            contentDescription = "返回"
-                        )
-                    }
-                },
-                colors = TopAppBarDefaults.topAppBarColors(
-                    containerColor = MaterialTheme.colorScheme.background,
-                    titleContentColor = MaterialTheme.colorScheme.onBackground
-                )
-            )
-        }
-    ) { padding ->
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(padding)
-        ) {
-            Surface(
-                modifier = Modifier.fillMaxWidth(),
-                color = MaterialTheme.colorScheme.surface,
-                tonalElevation = 2.dp
-            ) {
-                Row(
-                    modifier = Modifier.padding(16.dp),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Icon(
-                        Icons.Default.Face,
-                        contentDescription = null,
-                        tint = MaterialTheme.colorScheme.primary,
-                        modifier = Modifier.size(32.dp)
-                    )
-                    Spacer(modifier = Modifier.width(12.dp))
-                    Column {
-                        Text(
-                            text = "恶搞兔表情包",
-                            fontSize = 18.sp,
-                            fontWeight = FontWeight.Bold
-                        )
-                        Text(
-                            text = "共 ${emojiFiles.size} 个表情 · 点击查看",
-                            fontSize = 14.sp,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant
-                        )
-                    }
-                }
-            }
-            
-            if (isLoading) {
-                Box(
-                    modifier = Modifier.fillMaxSize(),
-                    contentAlignment = Alignment.Center
-                ) {
-                    CircularProgressIndicator()
-                }
-            } else {
-                LazyVerticalGrid(
-                    columns = GridCells.Fixed(2),
-                    contentPadding = PaddingValues(16.dp),
-                    horizontalArrangement = Arrangement.spacedBy(12.dp),
-                    verticalArrangement = Arrangement.spacedBy(12.dp)
-                ) {
-                    items(emojiFiles) { file ->
-                        EmojiItemCard(file = file)
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun EmojiItemCard(
-    file: File
-) {
-    Surface(
-        modifier = Modifier.aspectRatio(1f),
-        shape = MaterialTheme.shapes.medium,
-        tonalElevation = 2.dp,
-        onClick = { }
-    ) {
-        Box(
-            modifier = Modifier.fillMaxSize()
-        ) {
-            AsyncImage(
-                model = ImageRequest.Builder(LocalContext.current)
-                    .data(file)
-                    .crossfade(true)
-                    .build(),
-                contentDescription = file.nameWithoutExtension,
-                modifier = Modifier.fillMaxSize(),
-                contentScale = ContentScale.Fit
-            )
-            
-            Surface(
-                modifier = Modifier.align(Alignment.BottomCenter),
-                color = MaterialTheme.colorScheme.surface.copy(alpha = 0.8f)
-            ) {
-                Text(
-                    text = file.nameWithoutExtension,
-                    modifier = Modifier.padding(horizontal = 8.dp, vertical = 4.dp),
-                    fontSize = 12.sp,
-                    fontWeight = FontWeight.Medium,
-                    maxLines = 1
-                )
-            }
-        }
-    }
-}
