+++ plugins/kaomoji/src/main/java/com/kingzcheung/kime/plugin/kaomoji/KaomojiPlugin.kt
@@ -42,19 +42,4 @@ class KaomojiPlugin : EmojiPlugin {
     }
     
     override suspend fun getCategories(): List<String> = listOf("颜文字")
-    
-    override fun hasSettings(): Boolean = true
-    
-    override fun openSettings(context: Context) {
-        try {
-            val intent = android.content.Intent()
-            intent.setClassName(
-                "com.kingzcheung.kime.plugin.kaomoji",
-                "com.kingzcheung.kime.plugin.kaomoji.PluginSettingsActivity"
-            )
-            context.startActivity(intent)
-        } catch (e: Exception) {
-            Toast.makeText(context, "无法打开设置: ${e.message}", Toast.LENGTH_SHORT).show()
-        }
-    }
 }
